Load CEBL Coaches

A function to load the Canadian Elite Basketball League (CEBL) coaches.

Usage

load_cebl_coaches(seasons = NULL)

Arguments

seasons

A vector of seasons (years) to filter the data.

Value

A data frame containing the CEBL coaches for the specified seasons.

Column Name	Type

game_id	dbl

season	dbl

team_name	chr

coach_name	chr

coach_type	chr

first_name	chr

first_name_initial	chr

last_name	chr

last_name_initial	chr

international_first_name	chr

international_first_name_initial	chr

international_last_name	chr

international_last_name_initial	chr

scoreboard_name	chr

Examples

load_cebl_coaches(2020:2021)
#> Rows: 2667 Columns: 14
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (12): team_name, coach_name, coach_type, first_name, first_name_initial,...
#> dbl  (2): game_id, season
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 403 × 14
#>    game_id season team_name  coach_name coach_type first_name first_name_initial
#>      <dbl>  <dbl> <chr>      <chr>      <chr>      <chr>      <chr>             
#>  1 1716674   2020 Niagara R… Victor  R… Head Coach Victor     V                 
#>  2 1716674   2020 Niagara R… Danielle … Assistant… Danielle   D                 
#>  3 1716674   2020 Niagara R… Mick Bett  Assistant… Mick       M                 
#>  4 1716674   2020 Hamilton … Ryan Schm… Head Coach Ryan       R                 
#>  5 1716674   2020 Hamilton … Chantal V… Assistant… Chantal    C                 
#>  6 1716674   2020 Hamilton … Sheldon C… Assistant… Sheldon    S                 
#>  7 1716676   2020 Ottawa Bl… Osvaldo J… Head Coach Osvaldo    O                 
#>  8 1716676   2020 Ottawa Bl… Jamie Cam… Assistant… Jamie      J                 
#>  9 1716676   2020 Ottawa Bl… Jafeth Ma… Assistant… Jafeth     J                 
#> 10 1716676   2020 Guelph Ni… Charles K… Head Coach Charles    C                 
#> # ℹ 393 more rows
#> # ℹ 7 more variables: last_name <chr>, last_name_initial <chr>,
#> #   international_first_name <chr>, international_first_name_initial <chr>,
#> #   international_last_name <chr>, international_last_name_initial <chr>,
#> #   scoreboard_name <chr>