Load CEBL Game Schedule

A function to load the Canadian Elite Basketball League (CEBL) schedule.

Usage

load_cebl_schedule(seasons = NULL)

Arguments

seasons

A vector of seasons (years) to filter the data. if NULL, defaults to the range from 2019 to the current year.

Value

A data frame containing the CEBL schedule for the specified seasons.

Column Name	Type

fiba_id	dbl

season	dbl

start_time_utc	dttm

status	chr

competition	chr

venue_name	chr

period	dbl

home_team_id	dbl

home_team_name	chr

home_team_score	dbl

home_team_logo_url	chr

home_team_url_stats_en	chr

home_team_url_stats_fr	chr

away_team_id	dbl

away_team_name	chr

away_team_score	dbl

away_team_logo_url	chr

away_team_url_stats_en	chr

away_team_url_stats_fr	chr

stats_url_en	chr

stats_url_fr	chr

cebl_stats_url_en	chr

cebl_stats_url_fr	chr

tickets_url_en	chr

tickets_url_fr	chr

id	dbl

fiba_json_url	chr

Examples

load_cebl_schedule(2020:2021)
#> Rows: 562 Columns: 27
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (18): status, competition, venue_name, home_team_name, home_team_logo_u...
#> dbl   (8): fiba_id, season, period, home_team_id, home_team_score, away_team...
#> dttm  (1): start_time_utc
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 80 × 27
#>    fiba_id season start_time_utc      status   competition venue_name     period
#>      <dbl>  <dbl> <dttm>              <chr>    <chr>       <chr>           <dbl>
#>  1 1716674   2020 2020-07-25 17:30:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  2 1716676   2020 2020-07-25 19:50:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  3 1716679   2020 2020-07-26 17:30:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  4 1716681   2020 2020-07-26 20:00:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  5 1716683   2020 2020-07-27 21:00:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  6 1716685   2020 2020-07-27 23:30:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  7 1716687   2020 2020-07-28 21:00:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  8 1716689   2020 2020-07-28 23:30:00 COMPLETE REGULAR     Meridian Cent…     NA
#>  9 1716691   2020 2020-07-29 23:00:00 COMPLETE REGULAR     Meridian Cent…     NA
#> 10 1716693   2020 2020-07-30 23:00:00 COMPLETE REGULAR     Meridian Cent…     NA
#> # ℹ 70 more rows
#> # ℹ 20 more variables: home_team_id <dbl>, home_team_name <chr>,
#> #   home_team_score <dbl>, home_team_logo_url <chr>,
#> #   home_team_stats_url_en <chr>, home_team_stats_url_fr <chr>,
#> #   away_team_id <dbl>, away_team_name <chr>, away_team_score <dbl>,
#> #   away_team_logo_url <chr>, away_team_stats_url_en <chr>,
#> #   away_team_stats_url_fr <chr>, stats_url_en <chr>, stats_url_fr <chr>, …