Load CEBL Team Boxscores

A function to load the Canadian Elite Basketball League (CEBL) team boxscores.

Usage

load_cebl_team_boxscores(seasons = NULL)

Arguments

seasons

A vector of seasons (years) to filter the data.

Value

A data frame containing the CEBL team boxscores for the specified seasons.

Column Name	Type

game_id	dbl

season	dbl

team_name	chr

short_name	chr

code	chr

team_score	dbl

minutes	chr

field_goals_made	dbl

field_goals_attempted	dbl

field_goal_percentage	dbl

two_point_field_goals_made	dbl

two_point_field_goals_attempted	dbl

two_point__percentage	dbl

three_point_field_goals_made	dbl

three_point_field_goals_attempted	dbl

three_point_percentage	dbl

free_throws_made	dbl

free_throws_attempted	dbl

free_throw_percentage	dbl

offensive_rebounds	dbl

defensive_rebounds	dbl

rebounds	dbl

assists	dbl

steals	dbl

turnovers	dbl

blocks	dbl

blocks_received	dbl

personal_fouls	dbl

fouls_drawn	dbl

total_fouls	dbl

bonus_fouls	dbl

points_in_the_paint	dbl

second_chance_points	dbl

points_from_turnovers	dbl

bench_points	dbl

fast_break_points	dbl

team_index_rating	dbl

team_index_rating_2	dbl

team_index_rating_3	dbl

team_index_rating_4	dbl

team_index_rating_5	dbl

team_index_rating_6	dbl

team_index_rating_7	dbl

team_fouls	dbl

team_turnovers	dbl

team_rebounds	dbl

team_defensive_rebounds	dbl

team_offensive_rebounds	dbl

period_1_score	dbl

period_2_score	dbl

period_3_score	dbl

period_4_score	dbl

biggest_lead	dbl

biggest_scoring_run	dbl

time_leading	dbl

lead_changes	dbl

times_scores_level	dbl

timeouts_left	dbl

head_coach	chr

assistant_coach_1	chr

assistant_coach_2	chr

international_team_name	chr

international_short_name	chr

international_code	chr

logo	chr

logo_t_url	chr

logo_t_size	chr

logo_t_height	dbl

logo_t_width	dbl

logo_t_bytes	dbl

logo_s_url	chr

logo_s_size	chr

logo_s_height	dbl

logo_s_width	dbl

logo_s_bytes	dbl

Examples

load_cebl_team_boxscores(2020:2021)
#> Rows: 1122 Columns: 75
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (14): team_name, short_name, code, minutes, head_coach, assistant_coach_...
#> dbl (60): game_id, season, team_score, field_goals_made, field_goals_attempt...
#> lgl  (1): logo
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 160 × 75
#>    game_id season team_name short_name code  team_score minutes field_goals_made
#>      <dbl>  <dbl> <chr>     <chr>      <chr>      <dbl> <chr>              <dbl>
#>  1 1716674   2020 Niagara … Niagara    NRL           97 200:00                36
#>  2 1716674   2020 Hamilton… Hamilton   HHB           85 200:00                27
#>  3 1716676   2020 Ottawa B… Blackjacks OBJ           71 200:00                24
#>  4 1716676   2020 Guelph N… Guelph     GNH           89 200:00                32
#>  5 1716679   2020 Saskatch… Saskatche… SKR           96 200:00                35
#>  6 1716679   2020 Niagara … Niagara    NRL           79 200:00                30
#>  7 1716681   2020 Edmonton… Edmonton   EDS          100 200:00                35
#>  8 1716681   2020 Fraser V… Fraser Va… FVB          113 200:00                40
#>  9 1716683   2020 Hamilton… Hamilton   HHB           97 200:00                33
#> 10 1716683   2020 Guelph N… Guelph     GNH           71 200:00                20
#> # ℹ 150 more rows
#> # ℹ 67 more variables: field_goals_attempted <dbl>,
#> #   field_goal_percentage <dbl>, two_point_field_goals_made <dbl>,
#> #   two_point_field_goals_attempted <dbl>, two_point_percentage <dbl>,
#> #   three_point_field_goals_made <dbl>,
#> #   three_point_field_goals_attempted <dbl>, three_point_percentage <dbl>,
#> #   free_throws_made <dbl>, free_throws_attempted <dbl>, …